# Gateway/Adapters/Http/Contract/customerbffopenapi.yaml
openapi: 3.0.0
info:
  title: Customer BFF
  version: 1.0.0
...
paths:
  /rent/start:
    post:
      operationId: startRent
      tags:
        - rent
      summary: Начать аренду
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartRentRequest'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartRentResponse'
        '500':
          description: Внутренняя ошибка сервера
  /rent/{rentId}/complete:
    post:
      operationId: completeRent
      tags:
        - rent
      summary: Завершить аренду
      parameters:
        - in: path
          name: rentId
          schema:
            type: string
            format: uuid
          required: true
          description: Идентификатор аренды
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteRentResponse'
        '404':
          description: Аренда с таким идентификатором не существует
        '500':
          description: Внутренняя ошибка сервера
  /rent/{rentId}/get-current-amount:
    get:
      operationId: getCurrentAmountForRent
      tags:
        - rent
      summary: Получить все текущие аренды
      parameters:
        - in: path
          name: rentId
          schema:
            type: string
            format: uuid
          required: true
          description: Идентификатор аренды
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCurrentAmountRentResponse'
        '500':
          description: Внутренняя ошибка сервера
  /driving-license/{licenseId}:
    get:
      operationId: getDrivingLicenseById
      tags:
        - driving-license
      summary: Получение водительского удостоверения по идентификатору
      parameters:
        - in: path
          name: licenseId
          schema:
            type: string
            format: uuid
          required: true
          description: Идентификатор водительского удостоверения
      responses:
        '200':
          description: Успешная операция
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLicenseByIdResponse'
        '500':
          description: Внутренняя ошибка сервера
  /driving-license/upload:
    post:
      operationId: uploadDrivingLicense
      tags:
        - driving-license
      summary: Загрузка водительского удостоверения
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadLicenseRequest'
      responses:
        '200':
          description: Успешная операция
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadLicenseResponse'
        '500':
          description: Внутренняя ошибка сервера
  /driving-license/{licenseId}/upload-photos:
    post:
      operationId: uploadDrivingLicensePhotos
      tags:
        - driving-license
      summary: Загрузка фотографий водительского удостоверения
      parameters:
        - in: path
          name: licenseId
          schema:
            type: string
            format: uuid
          required: true
          description: Идентификатор аренды
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadPhotosRequest'
      responses:
        '200':
          description: Успешная операция
        '500':
          description: Внутренняя ошибка сервера
...
  /identity/authenticate:
    post:
      operationId: authenticate
      tags:
        - identity
      summary: Аутентификация аккаунта
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticateRequest'
      responses:
        '200':
          description: Успешная операция
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticateResponse'
        '500':
          description: Внутренняя ошибка сервера
        '404':
          description: Аккаунт с таким идентификатором не найден
  /identity/account/create:
    post:
      operationId: createAccount
      tags:
        - identity
      summary: Создание аккаунта
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequest'
      responses:
        '201':
          description: Аккаунт успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResponse'
        '500':
          description: Внутренняя ошибка сервера
  /identity/{accountId}/confirmation-token/{confirmationToken}/confirm:
    post:
      operationId: confirmAccountEmail
      tags:
        - identity
      summary: Подтверждение email аккаунта
      parameters:
        - name: accountId
          in: path
          description: Идентификатор аккаунта
          required: true
          schema:
            type: string
            format: uuid
        - name: confirmationToken
          in: path
          description: Токен подтверждения
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmEmailRequest'
      responses:
        '200':
          description: Успешная операция
        '500':
          description: Внутренняя ошибка сервера
        '404':
          description: Токен подтверждения не найден
  /identity/refresh-access-token:
    post:
      operationId: refreshAccessToken
      tags:
        - identity
      summary: Обновление токена доступа
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshAccessTokenRequest'
      responses:
        '200':
          description: Успешная операция
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshAccessTokenResponse'
        '500':
          description: Внутренняя ошибка сервера
        '404':
          description: Refresh token с таким идентификатором не найден
  /identity/account/{accountId}/recover-password:
    post:
      operationId: recoverPassword
      tags:
        - identity
      summary: Восстановление пароля
      parameters:
        - name: accountId
          in: path
          description: Идентификатор аккаунта
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoverPasswordRequest'
      responses:
        '200':
          description: Успешная операция
        '500':
          description: Внутренняя ошибка сервера
        '404':
          description: Аккаунт с таким идентификатором не найден
  /identity/account/{accountId}/update-password:
    post:
      operationId: updatePassword
      tags:
        - identity
      summary: Обновление пароля
      parameters:
        - name: accountId
          in: path
          description: Идентификатор аккаунта
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '200':
          description: Успешная операция
        '500':
          description: Внутренняя ошибка сервера
        '404':
          description: Аккаунт с таким идентификатором не найден
  /model/{page}/{pageSize}:
    get:
      operationId: getAllModels
      tags:
        - model
      summary: Получение всех моделей автомобилей
      parameters:
        - name: page
          in: path
          description: Номер страницы
          required: true
          schema:
            type: integer
        - name: pageSize
          in: path
          description: Количество элементов на странице
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешная операция
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllModelsResponse'
        '500':
          description: Внутренняя ошибка сервера
  /model/{modelId}:
    get:
      operationId: getModelById
      tags:
        - model
      summary: Получение модели автомобиля по идентификатору
      parameters:
        - name: modelId
          in: path
          description: Идентификатор модели автомобиля
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успешная операция
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetModelByIdResponse'
        '500':
          description: Внутренняя ошибка сервера
...
components:
  schemas:
...
    AuthenticateRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          example: "example@example.com"
        password:
          type: string
          example: "password"
    AuthenticateResponse:
      type: object
      required:
        - accessToken
        - refreshToken
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
    CreateRequest:
      type: object
      required:
        - role
        - email
        - phone
        - password
      properties:
        email:
          type: string
          example: "example@example.com"
        phone:
          type: string
          example: "+79005008010"
        password:
          type: string
          example: "password"
    CreateResponse:
      type: object
      required:
        - accountId
      properties:
        accountId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
    ConfirmEmailRequest:
      type: object
      required:
        - confirmationToken
      properties:
        confirmationToken:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
    RefreshAccessTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
    RefreshAccessTokenResponse:
      type: object
      required:
        - accessToken
        - refreshToken
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
    RecoverPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          example: "example@example.com"
    UpdatePasswordRequest:
      type: object
      required:
        - newPassword
        - email
        - resetToken
      properties:
        newPassword:
          type: string
          example: "new-password"
        email:
          type: string
          example: "example@example.com"
        resetToken:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
...
    GetAllModelsResponse:
      type: object
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/ModelShortView'
    ModelShortView:
      type: object
      properties:
        model_id:
          type: string
        brand:
          type: string
        carModelId:
          type: string
    GetModelByIdResponse:
      type: object
      properties:
        model_id:
          type: string
        brand:
          type: string
        carModelId:
          type: string
        category:
          type: string
        pricePerMinute:
          type: number
        pricePerHour:
          type: number
        price_per_day:
          type: number
...
    GetLicenseByIdResponse:
      type: object
      required:
        - id
        - accountId
        - status
        - categories
        - number
        - name
        - cityOfBirth
        - dateOfBirth
        - codeOfIssue
        - dateOfIssue
        - dateOfExpiry
      properties:
        id:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/DrivingLicenseStatus'
        categories:
          type: array
          items:
            type: string
            example: "B"
        number:
          type: string
        name:
          type: string
        cityOfBirth:
          type: string
        dateOfBirth:
          type: string
          format: date
        codeOfIssue:
          type: string
        dateOfIssue:
          type: string
          format: date
        dateOfExpiry:
          type: string
          format: date
        frontPhotoS3Url:
          type: string
        backPhotoS3Url:
          type: string
    UploadLicenseRequest:
      type: object
      required:
        - accountId
        - categories
        - number
        - firstName
        - lastName
        - cityOfBirth
        - dateOfBirth
        - codeOfIssue
        - dateOfIssue
        - dateOfExpiry
      properties:
        accountId:
          type: string
          format: uuid
        categories:
          type: array
          items:
            type: string
            example: "B"
        number:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        patronymic:
          type: string
        cityOfBirth:
          type: string
        dateOfBirth:
          type: string
          format: date
        codeOfIssue:
          type: string
        dateOfIssue:
          type: string
          format: date
        dateOfExpiry:
          type: string
          format: date
    UploadLicenseResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
    UploadPhotosRequest:
      type: object
      required:
        - frontPhotoBytes
        - backPhotoBytes
      properties:
        frontPhotoBytes:
          type: string
          format: byte
        backPhotoBytes:
          type: string
          format: byte
...
    DrivingLicenseStatus:
      type: string
      enum:
        - approved
        - pendingPhotosAdding
        - pendingProcessing
        - rejected
        - expired
    StartRentRequest:
      type: object
      properties:
        customerId:
          type: string
          description: Идентификатор клиента
        bookingId:
          type: string
          description: Идентификатор бронирования
        vehicleId:
          type: string
          description: Идентификатор транспортного средства
    StartRentResponse:
      type: object
      properties:
        rentId:
          type: string
          description: Идентификатор аренды
    CompleteRentResponse:
      type: object
      properties:
        actual_amount:
          type: number
          format: double
          description: Фактическая сумма аренды
...
    GetCurrentAmountRentResponse:
      type: object
      properties:
        rentId:
          type: string
          description: Идентификатор аренды
        current_amount:
          type: number
          format: double
          description: Текущая сумма аренды
...